+++ consul/consul.go
@@ -104,10 +104,8 @@ func (r *ConsulAdapter) buildCheck(service *bridge.Service) *consulapi.AgentServ
 		if timeout := service.Attrs["check_timeout"]; timeout != "" {
 			check.Timeout = timeout
 		}
-		//} else if cmd := service.Attrs["check_cmd"]; cmd != "" {
-		//	check.Script = fmt.Sprintf("check-cmd %s %s %s", service.Origin.ContainerID[:12], service.Origin.ExposedPort, cmd)
-		//} else if script := service.Attrs["check_script"]; script != "" {
-		//	check.Script = r.interpolateService(script, service)
+	} else if script := service.Attrs["check_script"]; script != "" {
+		check.Args = []string{r.interpolateService(script, service)}
 	} else if ttl := service.Attrs["check_ttl"]; ttl != "" {
 		check.TTL = ttl
 	} else if tcp := service.Attrs["check_tcp"]; tcp != "" {
@@ -118,16 +116,15 @@ func (r *ConsulAdapter) buildCheck(service *bridge.Service) *consulapi.AgentServ
 	} else {
 		return nil
 	}
-	//if check.Script != "" || check.HTTP != "" || check.TCP != "" {
-	if check.HTTP != "" || check.TCP != "" {
+	if len(check.Args) > 0 || check.HTTP != "" || check.TCP != "" {
 		if interval := service.Attrs["check_interval"]; interval != "" {
 			check.Interval = interval
 		} else {
 			check.Interval = DefaultInterval
 		}
 	}
-	if deregister_after := service.Attrs["check_deregister_after"]; deregister_after != "" {
-		check.DeregisterCriticalServiceAfter = deregister_after
+	if deregisterAfter := service.Attrs["check_deregister_after"]; deregisterAfter != "" {
+		check.DeregisterCriticalServiceAfter = deregisterAfter
 	}
 	return check
 }